tests: split test_loaders in multiple files
