+++ tests/tests_unit/test_cdf_tk/test_loaders/test_data_model_loader.py
@@ -1,0 +1,54 @@
+from unittest.mock import MagicMock
+
+from cognite.client.data_classes import data_modeling as dm
+
+from cognite_toolkit._cdf_tk.commands import DeployCommand
+from cognite_toolkit._cdf_tk.loaders import DataModelLoader
+from cognite_toolkit._cdf_tk.utils import CDFToolConfig
+from tests.tests_unit.approval_client import ApprovalCogniteClient
+
+
+class TestDataModelLoader:
+    def test_update_data_model_random_view_order(self, cognite_client_approval: ApprovalCogniteClient):
+        cdf_tool = MagicMock(spec=CDFToolConfig)
+        cdf_tool.verify_client.return_value = cognite_client_approval.mock_client
+        cdf_tool.verify_capabilities.return_value = cognite_client_approval.mock_client
+        cdf_tool.client = cognite_client_approval.mock_client
+        cdf_data_model = dm.DataModel(
+            space="sp_space",
+            external_id="my_model",
+            version="1",
+            views=[
+                dm.ViewId(space="sp_space", external_id="first", version="1"),
+                dm.ViewId(space="sp_space", external_id="second", version="1"),
+            ],
+            last_updated_time=1,
+            created_time=1,
+            description=None,
+            name=None,
+            is_global=False,
+        )
+        # Simulating that the data model is available in CDF
+        cognite_client_approval.append(dm.DataModel, cdf_data_model)
+
+        local_data_model = dm.DataModelApply(
+            space="sp_space",
+            external_id="my_model",
+            version="1",
+            views=[
+                dm.ViewId(space="sp_space", external_id="second", version="1"),
+                dm.ViewId(space="sp_space", external_id="first", version="1"),
+            ],
+            description=None,
+            name=None,
+        )
+
+        loader = DataModelLoader.create_loader(cdf_tool, None)
+        cmd = DeployCommand(print_warning=False)
+        to_create, to_change, unchanged = cmd.to_create_changed_unchanged_triple(
+            dm.DataModelApplyList([local_data_model]), loader
+        )
+
+        assert len(to_create) == 0
+        assert len(to_change) == 0
+        assert len(unchanged) == 1
